feat: change file name, improving code readability
